Adding The Files For Automatic SetUp Of Array Runs
There are so many files in this commit because these files were being
used for a long time, but for some stupid reason I did not included in
the repo. Well, not really a stupid reason! The reason was that I did
not know how to use git. But now I know a bit more and decided to
include the rest of the files that should have been there from the very
beginning.

++0.collectorBoundaries:
-Within the directory, the boundary conditions
for the many collectors are set in this files, and later this files
are called from the boundary condition section of the flow fields.
Obviously, these files need first to be copied into the simulation case
directory.

++system.FlowFields:
-Within the directory, several dictionaries are set, and these
dictionaries are copied into the case directory (within system) and
modified there to be adapted to the simulation specific parameters (like
Reynolds number or endtime).

++system.ForcesSampling:
-Most of these files were created with a matlab tool that generates the
sampling points to be indicated into OpenFOAM.
-The "samplingRings*.txt" files are copied into the case directory
(within system) and then ingested into the sampleDict of the
case.Sampled points are positioned in rings around the collectors.
-The "samplingLines*.txt" work the same, but points are positioned in
lines along the domain.
-The different particle sizes to be analyzed (sampled within the
simulation) are setup in the files named "*Rp*.txt". These files are
copied to the case directory and read by the Python analyzer.
-The "forces*" files are set up for the analysis of forces around the
cylinders and are to be ingested in the controDict.
-The "generalGeometry*" are used somewhere to define the type of
simulation
-The "additional*" files were defined with that name because came from
an idea on top of the original analysis that was being performed and
therefore the name.